PR feedback
Removed internal declaration of BindingFlags - relying on the one from framework.

Added internal declaration for the new attributes with the expectation that we're going to remove them (ifdef) once we move to the framework which already has them.
