+++ src/linker/Linker.Dataflow/AggregateFlowAnnotationSource.cs
@@ -2,9 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Mono.Cecil;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using Mono.Cecil;
 
 namespace Mono.Linker.Dataflow
 {
