+++ src/linker/Linker.Dataflow/AttributeFlowAnnotationSource.cs
@@ -2,8 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Mono.Cecil;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using Mono.Cecil;
 
 namespace Mono.Linker.Dataflow
 {
